+++ oxide-auth-actix/examples/actix-example/src/main.rs
@@ -130,13 +130,12 @@ pub fn main() {
 
 impl State {
     pub fn preconfigured() -> Self {
-        let mut registrar = ClientMap::new();
-        registrar.set_password_policy(Pbkdf2::default());
+        let mut registrar = ClientMap::new(Pbkdf2::default());
         registrar.register_client(Client::public(
             "LocalClient",
             "http://localhost:8021/endpoint".parse().unwrap(),
             "default-scope".parse().unwrap(),
-        )).unwrap();
+        ));
 
         State {
             endpoint: Generic {
